autotests/pyproject.toml
[tool.poetry]
name = "onlinetlabs-autotests"
version = "0.1.0"
description = "API-автотесты для onlinetlabs backend"
authors = []
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
httpx = "^0.28"
pytest = "^8.0"
pytest-asyncio = "^0.25"
pydantic = "^2.0"
python-dotenv = "^1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
